Refactor Docker Compose infrastructure by moving DockerComposeEnvironmentContext to a new file and cleaning up unused imports
